+++ osrframework/phonefy.py
@@ -3,7 +3,7 @@
 #
 ##################################################################################
 #
-#    Copyright 2016 Félix Brezo and Yaiza Rubio (i3visio, contacto@i3visio.com)
+#    Copyright 2015-2017 Félix Brezo and Yaiza Rubio (i3visio, contacto@i3visio.com)
 #
 #    This program is part of OSRFramework. You can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -21,15 +21,15 @@
 ##################################################################################
 
 '''
-phonefy.py Copyright (C) F. Brezo and Y. Rubio (i3visio) 2016
+phonefy.py Copyright (C) F. Brezo and Y. Rubio (i3visio) 2015-2017
 This program comes with ABSOLUTELY NO WARRANTY.
 This is free software, and you are welcome to redistribute it under certain conditions.  For additional info, visit to <http://www.gnu.org/licenses/gpl-3.0.txt>.
 '''
 __author__ = "Felix Brezo, Yaiza Rubio "
-__copyright__ = "Copyright 2016, i3visio"
+__copyright__ = "Copyright 2015-2017, i3visio"
 __credits__ = ["Felix Brezo", "Yaiza Rubio"]
 __license__ = "GPLv3+"
-__version__ = "v2.0"
+__version__ = "v2.1"
 __maintainer__ = "Felix Brezo, Yaiza Rubio"
 __email__ = "contacto@i3visio.com"
 
@@ -115,6 +115,8 @@ This is free software, and you are welcome to redistribute it under certain cond
         print "Then, place an issue in the Github project: <https://github.com/i3visio/osrframework/issues>."
         print "Note that otherwise, we won't know about it!"
         print
+
+    return results
 
 def getParser():
     import osrframework.utils.configuration as configuration